feat: refactor job section components and add tech stack section with image
assets:
Added the BMW svg for experience section.

JobCard:
Added a image for the header of each job so it shows title and image of
the company

Job data:
Extracted job data to its own file for easier management

Tech Stack Section:
Added a tech stack section to its own component for the use of making
easier to manage later on.

App.ts:
Added the tech stack section that was taken out.